+++ src/Preferences.js
@@ -8,7 +8,7 @@ define(function (require, exports, module) {
 
     // Defining preferences
     prefs.definePreference("itemsWithBibtexKeyOnly", "boolean", true);
-    prefs.definePreference("extraFieldPrefix",       "string",  "bibtex");
+    prefs.definePreference("extraFieldPrefix",       "string",  "bibtex:");
     prefs.definePreference("docTypesToScan",         "string",  "md txt");
     prefs.definePreference("bibFolderName",          "string",  "brackets-zotero-bibtex-DB");
     prefs.definePreference("scanInterval",           "number",  10000);
@@ -16,5 +16,20 @@ define(function (require, exports, module) {
 
     // Saving preferences
     prefs.save();
+    
+    // Get preference
+    function get(name) {
+        return prefs.get(name);
+    }
+        
+    // Set preference
+    function set(name, value) {
+        prefs.set(name, value);
+    }
+    
+    module.exports = {
+        get: get,
+        set: set
+    }
 
 });